[MS] Replaced the rest of components/core by megashark-lib
