--- client/src/components/core/index.ts
@@ -1,10 +1,0 @@
-// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
-
-// ** ms-types ** //
-export * from '@/components/core/ms-types';
-
-// ** ms-modal ** //
-export * from '@/components/core/ms-modal';
-
-// ** ms-tooltip ** //
-export * from '@/components/core/ms-tooltip';
