--- client/src/components/core/ms-modal/MsQuestionModal.vue
@@ -1,48 +1,0 @@
-<!-- Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS -->
-
-<template>
-  <ms-modal
-    :title="title"
-    :subtitle="subtitle"
-    :close-button="{ visible: false }"
-    :cancel-button="{
-      disabled: false,
-      label: noText || 'QuestionModal.no',
-      onClick: onNo,
-    }"
-    :confirm-button="{
-      disabled: false,
-      label: yesText || 'QuestionModal.yes',
-      onClick: onYes,
-      theme: yesIsDangerous ? MsReportTheme.Error : undefined,
-    }"
-    @on-enter-keyup="onYes"
-  />
-</template>
-
-<script setup lang="ts">
-import MsModal from '@/components/core/ms-modal/MsModal.vue';
-import { MsModalResult } from '@/components/core/ms-modal/types';
-import { MsReportTheme } from '@/components/core/ms-types';
-import { Translatable } from 'megashark-lib';
-import { modalController } from '@ionic/vue';
-
-defineProps<{
-  title: Translatable;
-  subtitle: Translatable;
-  yesText?: Translatable;
-  noText?: Translatable;
-  yesIsDangerous?: boolean;
-}>();
-
-async function onYes(): Promise<boolean> {
-  const res = await modalController.dismiss(null, MsModalResult.Confirm);
-  return res;
-}
-
-async function onNo(): Promise<boolean> {
-  return await modalController.dismiss(null, MsModalResult.Cancel);
-}
-</script>
-
-<style lang="scss" scoped></style>
