--- client/src/components/core/ms-modal/index.ts
@@ -1,7 +1,0 @@
-// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
-import MsAlertModal from '@/components/core/ms-modal/MsAlertModal.vue';
-import MsModal from '@/components/core/ms-modal/MsModal.vue';
-
-export * from '@/components/core/ms-modal/types';
-export * from '@/components/core/ms-modal/utils';
-export { MsAlertModal, MsModal };
