--- client/src/components/core/ms-tooltip/MsInformationTooltip.vue
@@ -1,32 +1,0 @@
-<!-- Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS -->
-
-<template>
-  <ion-icon
-    :slot="slot"
-    :icon="informationCircle"
-    class="icon"
-    @click="openInformationTooltip($event, text)"
-  />
-</template>
-
-<script setup lang="ts">
-import { openInformationTooltip } from '@/components/core/ms-tooltip/utils';
-import { Translatable } from 'megashark-lib';
-import { IonIcon } from '@ionic/vue';
-import { informationCircle } from 'ionicons/icons';
-
-defineProps<{
-  text: Translatable;
-  slot: 'start' | 'end';
-}>();
-</script>
-
-<style scoped lang="scss">
-.icon {
-  --color: var(--parsec-color-light-secondary-grey);
-
-  &:hover {
-    color: var(--parsec-color-light-secondary-text);
-  }
-}
-</style>
