--- client/src/components/core/ms-tooltip/index.ts
@@ -1,7 +1,0 @@
-// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
-
-export * from '@/components/core/ms-tooltip/utils';
-
-import MsInformationTooltip from '@/components/core/ms-tooltip/MsInformationTooltip.vue';
-
-export { MsInformationTooltip };
