+++ client/src/components/workspaces/WorkspaceUserRole.vue
@@ -20,8 +20,7 @@
 </template>
 
 <script setup lang="ts">
-import { MsAppearance, MsOptions } from '@/components/core';
-import { MsDropdown } from 'megashark-lib';
+import { MsAppearance, MsDropdown, MsOptions } from 'megashark-lib';
 import UserAvatarName from '@/components/users/UserAvatarName.vue';
 import { canChangeRole } from '@/components/workspaces/utils';
 import { UserProfile, UserTuple, WorkspaceRole } from '@/parsec';
