+++ client/src/services/informationManager.ts
@@ -1,12 +1,11 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
-import { MsAlertModal, MsAlertModalConfig, MsReportTheme } from '@/components/core';
 import { ConnectionHandle, EntryName, FsPath, SizeInt, UserID, WorkspaceHandle, WorkspaceRole } from '@/parsec';
 import { getConnectionHandle } from '@/router';
 import { NotificationManager } from '@/services/notificationManager';
 import { ToastManager } from '@/services/toastManager';
 import { modalController } from '@ionic/vue';
-import { Translatable } from 'megashark-lib';
+import { MsAlertModal, MsAlertModalConfig, MsReportTheme, Translatable } from 'megashark-lib';
 import { v4 as uuid4 } from 'uuid';
 
 export const InformationManagerKey = 'information';
