+++ client/src/views/devices/DevicesPage.vue
@@ -126,8 +126,7 @@
 </template>
 
 <script setup lang="ts">
-import { MsModalResult } from '@/components/core';
-import { MsImage, PasswordLock } from 'megashark-lib';
+import { MsImage, MsModalResult, PasswordLock } from 'megashark-lib';
 import DeviceCard from '@/components/devices/DeviceCard.vue';
 import { OwnDeviceInfo, hasRecoveryDevice, listOwnDevices } from '@/parsec';
 import { Routes, navigateTo } from '@/router';
