+++ client/src/views/files/FileDetailsModal.vue
@@ -97,8 +97,7 @@
 
 <script setup lang="ts">
 import { formatFileSize, getFileIcon, shortenFileName } from '@/common/file';
-import { MsModal } from '@/components/core';
-import { Clipboard, Folder, MsImage } from 'megashark-lib';
+import { Clipboard, Folder, MsImage, MsModal } from 'megashark-lib';
 import { EntryStat, EntryStatFile } from '@/parsec';
 import { IonButton, IonIcon, IonLabel, IonPage, IonText } from '@ionic/vue';
 import { cloudDone, cloudOffline, copy } from 'ionicons/icons';
