+++ client/src/views/users/UserDetailsModal.vue
@@ -83,8 +83,7 @@
 </template>
 
 <script setup lang="ts">
-import { formatTimeSince } from 'megashark-lib';
-import { MsModal } from '@/components/core';
+import { MsModal, formatTimeSince } from 'megashark-lib';
 import WorkspaceTagRole from '@/components/workspaces/WorkspaceTagRole.vue';
 import { SharedWithInfo, UserInfo, getWorkspacesSharedWith } from '@/parsec';
 import { Information, InformationLevel, InformationManager, PresentationMode } from '@/services/informationManager';
